move shared routine to utils.js, remove flop peak reduction
